Make temp-index persistence conditional on a system property (defaults to false, i.e. do-not-keep).

+++ mimir-test/src/gate/mimir/test/QueryTests.java
@@ -128,10 +128,14 @@ public class QueryTests {
    */
   @AfterClass
   public static void oneTimeTearDown() {
-    for(File indexDir : indexDirs) {
-      // recursively delete index dir
-      if(!TestUtils.deleteDir(indexDir)) {
-        System.err.println("Could not delete index directory " + indexDir);
+    boolean keepIndex = Boolean.parseBoolean(
+      System.getProperty("keepTestIndex", "false"));
+    if(!keepIndex) {
+      for(File indexDir : indexDirs) {
+        // recursively delete index dir
+        if(!TestUtils.deleteDir(indexDir)) {
+          System.err.println("Could not delete index directory " + indexDir);
+        }      
       }      
     }
   }
